Cleanup styles and fix question page styling :art:
Integrates and consolidates all the progress bar styling, adds support
for Source Sans Pro across much of the app, simplifies the question page
styling, comments numerous portions of the file, and brings the question
page more in line with the graphic design frames.

However, more is necessary to organize the remainder of the file and fix
other styling issues.

Furthermore, there's an issue where the progress bar lines may not quite
match up color-wise and in terms of width with the left and right
bracket images, though this may be an issue on my end. If you're using a
Mac, the screen is low resolution enough that you don't need high
resolution assets. However, if you're using a device with a crisp,
gorgeous, high resolution screen covering over 99% of the Adobe RGB
color spectrum (cough, Surface Pro 4), then some higher res images would
be useful to reach its full potential.