Update tie calculation on results page
First, each question in questions.json was given a unique priority in a
previous commit, starting a zero for the highest priority. For multiple
choice quizzes, if there's a tie between multiple categories, loop
through each question by priority until a tie breaker can be determined
between the tied categories, going to the next question of lesser
priority if the user selected an answer that wasn't one of the winning
categories for a given question. Also, add some miscellaneous changes to
the template. Fix various syntax errors.